renamed folders to follow standards
