+++ CityOfBinds/src/__init__.py
@@ -1,15 +1,15 @@
-from .Triggers.trigger import Trigger, WASDTrigger
-from .Triggers.mixin import TriggerMixin
-from .SlashCommands.slashcommand import SlashCommand
-from .SlashCommands.commandgroup import CommandGroup
-from .SlashCommands.commandstemplate import CommandsTemplate
-from .SlashCommands.power import Power
-from .Binds.bind import Bind, WASDBind
-from .Binds.bindtemplate import BindTemplate
-from .BindFile.comments import Comment, CommentBanner
-from .BindFile.bindfile import BindFile
-from .BindFile.bindfiletemplate import BindFileTemplate
-from .RotatingBinds.rotatingbind import RotatingBind
+from .triggers.trigger import Trigger, WASDTrigger
+from .triggers.mixin import TriggerMixin
+from .slash_commands.slashcommand import SlashCommand
+from .slash_commands.commandgroup import CommandGroup
+from .slash_commands.commandstemplate import CommandsTemplate
+from .slash_commands.power import Power
+from .binds.bind import Bind, WASDBind
+from .binds.bindtemplate import BindTemplate
+from .bind_file.comments import Comment, CommentBanner
+from .bind_file.bindfile import BindFile
+from .bind_file.bindfiletemplate import BindFileTemplate
+from .rotating_binds.rotatingbind import RotatingBind
 from ..utils.baseconverter import BaseConverter
 from ..utils.pathgenerator import PathGenerator
 from ..utils.Templates import templates as Templates
